ODB-287 Remove unnecessary and potentially harmful cast.

+++ src/odb_api/odbql.cc
@@ -744,14 +744,13 @@ int odbql_column_count(odbql_stmt *stmt)
 double odbql_value_double(odbql_value* vp)
 {
     typedef double* double_p; 
-    return * double_p(vp);
+    return *double_p(vp);
 }
 
 int odbql_value_int(odbql_value* vp)
 {
     typedef double* double_p; 
-    //return int( * double_p(vp));
-    return static_cast<unsigned long>( *double_p(vp));
+    return *double_p(vp);
 }
 
 } // extern "C"
